+++ src/connection/publisher.js
@@ -53,8 +53,8 @@ class ConnectionPublisher extends ConnectionBase {
         }
         if (typeof this._pc.ontrack === 'undefined') {
           this._pc.onaddstream = event => {
-            if (this.clientId !== event.stream.id) {
-              this.remoteClientIds.push(stream.id);
+            if (this.connectionId !== event.stream.id) {
+              this.remoteConnectionIds.push(stream.id);
               this._callbacks.addstream(event);
             }
           };
@@ -63,17 +63,17 @@ class ConnectionPublisher extends ConnectionBase {
             const stream = event.streams[0];
             if (!stream) return;
             if (stream.id === 'default') return;
-            if (stream.id === this.clientId) return;
-            if (-1 < this.remoteClientIds.indexOf(stream.id)) return;
+            if (stream.id === this.connectionId) return;
+            if (-1 < this.remoteConnectionIds.indexOf(stream.id)) return;
             event.stream = stream;
-            this.remoteClientIds.push(stream.id);
+            this.remoteConnectionIds.push(stream.id);
             this._callbacks.addstream(event);
           };
         }
         this._pc.onremovestream = event => {
-          const index = this.remoteClientIds.indexOf(event.stream.id);
+          const index = this.remoteConnectionIds.indexOf(event.stream.id);
           if (-1 < index) {
-            delete this.remoteClientIds[index];
+            delete this.remoteConnectionIds[index];
           }
           this._callbacks.removestream(event);
         };