+++ CHANGES.md
@@ -11,6 +11,11 @@
 
 ## develop
 
+## 1.12.0
+- [UPDATE] example の整理
+- [UPDATE] development build 時に sora-js-sdk の version に '-dev' をつけるようにする
+- [ADD] Signaling Option に client_id を追加する
+
 ## 1.11.0
 - [UPDATE] Safari の Unified Plan, Plan B 両方に対応する
 - [UPDATE] Simulcast option が使えるブラウザ判定を変更する